Kernels for tile based approach. Should set all pixels to 1

+++ Image.hpp
@@ -4,7 +4,7 @@
 
 #include "Pixel.hpp"
 
-const size_t IMAGE_SIZE = 1024;
+const int IMAGE_SIZE = 1024;
 using Image = std::array<std::array<Pixel, IMAGE_SIZE>, IMAGE_SIZE>;
 
 std::istream& operator>>(std::istream &is, Image &image);
